+++ forui/test/src/widgets/select_menu_tile/select_menu_tile_test.dart
@@ -102,6 +102,109 @@ void main() {
       expect(find.text('Item 1'), findsNothing);
       expect(find.text('Item 2'), findsNothing);
     });
+
+    testWidgets('callbacks called', (tester) async {
+      var changes = 0;
+      var selections = 0;
+      (int, bool)? selection;
+
+      await tester.pumpWidget(
+        TestScaffold.app(
+          child: FSelectMenuTile<int>(
+            autoHide: true,
+            selectController: controller,
+            title: const Text('Repeat'),
+            onChange: (_) => changes++,
+            onSelect: (value) {
+              selections++;
+              selection = value;
+            },
+            menu: [
+              FSelectTile(title: const Text('1'), value: 1),
+            ],
+          ),
+        ),
+      );
+      await tester.tap(find.byType(FSelectMenuTile<int>));
+      await tester.pumpAndSettle();
+
+      await tester.tap(find.text('1'));
+      await tester.pumpAndSettle();
+
+      expect(changes, 1);
+      expect(selections, 1);
+      expect(selection, (1, true));
+    });
+
+    testWidgets('update widget', (tester) async {
+      final controller = FSelectController<int>();
+
+      var firstChanges = 0;
+      var firstSelections = 0;
+      (int, bool)? firstSelection;
+
+      await tester.pumpWidget(
+        TestScaffold.app(
+          child: FSelectMenuTile<int>(
+            autoHide: true,
+            selectController: controller,
+            title: const Text('Repeat'),
+            onChange: (_) => firstChanges++,
+            onSelect: (value) {
+              firstSelections++;
+              firstSelection = value;
+            },
+            menu: [
+              FSelectTile(title: const Text('1'), value: 1),
+            ],
+          ),
+        ),
+      );
+      await tester.tap(find.byType(FSelectMenuTile<int>));
+      await tester.pumpAndSettle();
+
+      await tester.tap(find.text('1'));
+      await tester.pumpAndSettle();
+
+      expect(firstChanges, 1);
+      expect(firstSelections, 1);
+      expect(firstSelection, (1, true));
+
+      var secondChanges = 0;
+      var secondSelections = 0;
+      (int, bool)? secondSelection;
+
+      await tester.pumpWidget(
+        TestScaffold.app(
+          child: FSelectMenuTile<int>(
+            autoHide: true,
+            selectController: controller,
+            title: const Text('Repeat'),
+            onChange: (_) => secondChanges++,
+            onSelect: (value) {
+              secondSelections++;
+              secondSelection = value;
+            },
+            menu: [
+              FSelectTile(title: const Text('1'), value: 1),
+            ],
+          ),
+        ),
+      );
+      await tester.tap(find.byType(FSelectMenuTile<int>));
+      await tester.pumpAndSettle();
+
+      await tester.tap(find.text('1'));
+      await tester.pumpAndSettle();
+
+      expect(firstChanges, 1);
+      expect(firstSelections, 1);
+      expect(firstSelection, (1, true));
+
+      expect(secondChanges, 1);
+      expect(secondSelections, 1);
+      expect(secondSelection, (1, false));
+    });
   });
 
   tearDown(() => controller.dispose());